Move the method to IgniteSemaphore
