+++ modules/core/src/main/java/org/apache/ignite/IgniteSemaphore.java
@@ -17,6 +17,10 @@
 
 package org.apache.ignite;
 
+import org.apache.ignite.lang.IgniteCallable;
+import org.apache.ignite.lang.IgniteFuture;
+import org.apache.ignite.lang.IgniteInClosure;
+
 import java.io.Closeable;
 import java.util.concurrent.TimeUnit;
 
@@ -243,6 +247,16 @@ public interface IgniteSemaphore extends Closeable {
      * @throws IllegalArgumentException if {@code permits} is negative
      */
     public void acquire(int permits) throws IgniteInterruptedException;
+
+    /**
+     * Acquires the given semaphore, executes the given callable and schedules the release of permits asynchronously
+     *
+     * @param callable the callable to execute
+     * @param numPermits the number of permits to acquire
+     * @throws Exception if the callable throws an exception
+     */
+    public <T> IgniteFuture<T> acquireAndExecute(IgniteCallable<IgniteFuture<T>> callable,
+                                                 int numPermits) throws Exception;
 
     /**
      * Releases the given number of permits, returning them to the semaphore.
